# PERF-#6122: Extract Feather's metadata without reading the whole file