<!--
Thank you for your contribution!
Please review the contributing docs: https://modin.readthedocs.io/en/latest/development/contributing.html
if you have questions about contributing.
-->

## What do these changes do?

This PR rewrites retrieving feather's metadata from reading the whole file to an approach suggested in the related arrow's issue: https://issues.apache.org/jira/browse/ARROW-10344

For the reproducer below, I've measured the following performance on 112 cores Intel(R) Xeon(R) Gold 6238R CPU @ 2.20GHz

| | master | this PR |
|-|-|-|
|read_feather| 6.05s | 0.99s |

```python
import modin.pandas as pd
import numpy as np
import tempfile
from timeit import default_timer as timer

NROWS = 10_000_000
NCOLS = 32

with tempfile.NamedTemporaryFile() as file:
    pd.DataFrame({f"col{i}": np.arange(NROWS) for i in range(NCOLS)}).to_feather(file.name)

    t1 = timer()
    df = pd.read_feather(file.name)
    print("read_feather", timer() - t1)
```

- [x] first commit message and PR title follow format outlined [here](https://modin.readthedocs.io/en/latest/development/contributing.html#commit-message-formatting)
  > **_NOTE:_**  If you edit the PR title to match this format, you need to add another commit (even if it's empty) or amend your last commit for the CI job that checks the PR title to pick up the new PR title.
- [x] passes `flake8 modin/ asv_bench/benchmarks scripts/doc_checker.py`
- [x] passes `black --check modin/ asv_bench/benchmarks scripts/doc_checker.py`
- [x] signed commit with `git commit -s` <!-- you can amend your commit with a signature via `git commit -amend -s` -->
- [x] Resolves #6122  <!-- issue must be created for each patch -->
- [x] tests are passing
- [x] module layout described at `docs/development/architecture.rst` is up-to-date <!-- if you have added, renamed or removed files or directories please update the documentation accordingly -->